feat: 分类参数 -> 动态参数,静态属性展示

+++ src/components/goods/Cate.vue
@@ -131,6 +131,15 @@ export default {
       })
     },
     async deleteCate (id) {
+      const confirmReslut = await this.$confirm('是否删除该商品分类?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'waring'
+      }).catch(err => err)
+
+      if (confirmReslut !== 'confirm') {
+        return this.$message.info('已取消删除')
+      }
       const { data: res } = await this.$http.delete(`categories/${id}`)
       if (res.meta.status !== 200) {
         return this.$message.error('删除分类失败!')
